I have splitted the app components into three, app.js, header.js and main.js
