+++ dom-merge-conflict/tasks/buttons-and-counter/src/app.js
@@ -1,4 +1,7 @@
 //increments the number in a node's text
+import { header } from "./header";  
+import { main } from "./main";
+
 function increment(node) {
   let current = node.textContent;
   node.textContent = Number(current) + 1;
@@ -7,25 +10,8 @@ function increment(node) {
 export function App() {
   const body = document.createElement("body");
 
-  const header = document.createElement("header");
-  header.innerHTML = `
-        <h1>Number Counter</h1>
-        <p>A simple counter. Press increment to increase the count by one.</p>
-    `;
   body.appendChild(header);
-
-  const main = document.createElement("main");
-  main.innerHTML = `
-        <p id="counter" data-testid="counter">0</p>
-        <button id="increment">Increment</button>
-    `;
   body.appendChild(main);
-
-  const button = body.querySelector("#increment");
-  const counter = body.querySelector("#counter");
-  button.addEventListener("click", () => {
-    increment(counter);
-  });
 
   return body;
 }
