+++ dom-merge-conflict/tasks/buttons-and-counter/src/main.js
@@ -1,0 +1,11 @@
+export const main = document.createElement("main");
+  main.innerHTML = `
+        <p id="counter" data-testid="counter">0</p>
+        <button id="increment">Increment</button>
+    `;
+
+  const button = body.querySelector("#increment");
+  const counter = body.querySelector("#counter");
+  button.addEventListener("click", () => {
+    increment(counter);
+  });
